added edit and completed password Hashing

+++ finaltask/src/main/java/com/task/Controller/UserController.java
@@ -5,6 +5,7 @@ import java.util.List;
 import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -27,51 +28,54 @@ public class UserController {
         this.service = service;
     }
 
+    BCryptPasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
+
     @GetMapping("/")
     public String homepage() {
         System.out.println("Into Home Page");
         return "Login";
     }
 
-    @GetMapping("/Details/AddUser")
-    public String addaddFormmethod(@ModelAttribute User user) {
+    // @GetMapping("/Details/AddUser")
+    // public String addaddFormmethod(@ModelAttribute User user) {
 
-        return "redirect:/Details/General";
+    // return "redirect:/Details/General";
 
-    }
+    // }
 
-    @PostMapping("/users/message")
-    public String addmethod(@ModelAttribute User user, @RequestParam(required = false) Integer userId, Model model,
-            RedirectAttributes redirectAttributes) {
-        String error;
-        User userExists = service.checkUserByMailId(user.getEmailId());
+    @PostMapping("/users/add")
+    public String addmethod(@ModelAttribute User user, Model model, RedirectAttributes redirectAttributes) {
+        user.setPassword(passwordEncoder.encode(user.getPassword()));
+        System.out.println("Encoded user password: " + user.getPassword());
 
-        if (userExists == null) {
-            String msg = service.addUsers(user);
-            model.addAttribute("msg", msg);
+        boolean isAdded = service.addUsers(user);
+
+        if (isAdded) {
+
+            model.addAttribute("msg", "User \"" + user.getUserName() + "\" Created Successfully!");
             return "message";
-
         } else {
-            error = "User already exists with this MailId!";
-            redirectAttributes.addFlashAttribute("message", error);
-            return "redirect:/Details/Signup";
+            redirectAttributes.addFlashAttribute("message", "User already exists with this MailId!");
+            return "redirect:/users/form?userId=";
         }
 
     }
+    // passwordEncoder.matches(password, userExists.getPassword())
 
     @PostMapping("/users")
     public String loginUser(@RequestParam String emailId, @RequestParam String password, HttpSession session,
-                            RedirectAttributes redirectAttributes) {
-    
+            RedirectAttributes redirectAttributes) {
+
         User userExists = service.checkUserByMailId(emailId);
         String msg;
-    
-        if (userExists != null && userExists.getPassword().equals(password)) {
+
+        if (userExists != null && passwordEncoder.matches(password, userExists.getPassword())) {
             service.updateCredentials(userExists);
             session.setAttribute("LoginUser", userExists);
-    
+            session.setAttribute("loggedInUserId", userExists.getUserId());
+
             List<User> users = service.fetchAllUsers();
-            session.setAttribute("UserList", users); 
+            session.setAttribute("UserList", users);
             return "redirect:/users";
         } else {
             msg = "Invalid email or password!";
@@ -79,42 +83,19 @@ public class UserController {
             return "redirect:/";
         }
     }
-    
+
     @GetMapping("/users")
-public String showUserPage(HttpSession session, Model model) {
-    // Fetch the latest UserList from the database instead of relying on the session
-    List<User> updatedUsers = service.fetchAllUsers();
-    model.addAttribute("UserList", updatedUsers);
-    model.addAttribute("loggedUser", session.getAttribute("LoginUser"));
-
-    return "Details";
-}
-
-
-
-    @GetMapping("/Details")
-
-    public String redirect(HttpSession session) {
-        if (session.getAttribute("user") == null) {
+    public String showUserPage(HttpSession session, Model model) {
+        if (session.getAttribute("LoginUser") == null) {
             return "redirect:/";
 
         }
+        List<User> updatedUsers = service.fetchAllUsers();
+        model.addAttribute("UserList", updatedUsers);
+        model.addAttribute("loggedUser", session.getAttribute("LoginUser"));
 
         return "Details";
-
     }
-
-    // @GetMapping("/users/form")
-
-    // public String redirectToSignUp(HttpSession session) {
-    // if (session.getAttribute("LoginUser") == null) {
-    // return "redirect:/";
-
-    // }
-
-    // return "General";
-
-    // }
 
     @GetMapping("/logout")
     public String logout(HttpSession session) {
@@ -122,10 +103,15 @@ public String showUserPage(HttpSession session, Model model) {
         return "redirect:/";
     }
 
-    @PostMapping("/users/viewInfo")
+    @GetMapping("/users/viewInfo")
 
     public String viewinfos(@RequestParam String userId, @RequestParam String employeeId, Model model,
             HttpSession session) {
+
+        if (session.getAttribute("LoginUser") == null) {
+            return "redirect:/";
+
+        }
 
         int user_id = Integer.parseInt(userId);
 
@@ -140,6 +126,10 @@ public String showUserPage(HttpSession session, Model model) {
 
     public String deleteUser(@RequestParam String userId, Model model, HttpSession session) {
 
+        if (session.getAttribute("LoginUser") == null) {
+            return "redirect:/";
+
+        }
         int user_id = Integer.parseInt(userId);
         service.deleteUserById(user_id);
 
@@ -147,27 +137,57 @@ public String showUserPage(HttpSession session, Model model) {
     }
 
     @GetMapping("/users/form")
-    public String addOrUpdate(@RequestParam String userId, Model model) {
+    public String addOrUpdate(@RequestParam String userId, Model model, HttpSession session) {
+
+        if (session.getAttribute("LoginUser") == null) {
+            return "redirect:/";
+
+        }
         if (userId.isEmpty() || userId == null) {
             model.addAttribute("user", null);
             return "General";
         } else {
             int userIdForAction = Integer.parseInt(userId);
             User userToUpdate = service.findUserById(userIdForAction);
+            model.addAttribute("LoggedInId", session.getAttribute("loggedInUserId"));
             model.addAttribute("user", userToUpdate);
             return "General";
         }
     }
 
-    @PostMapping("/user/update")
+    @PostMapping("/users/update")
     public String updatemethod(@ModelAttribute User updateUser, @RequestParam String refUserID, Model model,
-            RedirectAttributes redirectAttributes) {
+            RedirectAttributes redirectAttributes, HttpSession session) {
+
+        if (session.getAttribute("LoginUser") == null) {
+            return "redirect:/";
+
+        }
         int paramId = Integer.parseInt(refUserID);
-        String msg = service.updateUsers(updateUser, paramId);
 
-        model.addAttribute("Message", msg);
+        if (service.updateUsers(updateUser, paramId)) {
+            String msg = "User updated successfully!";
+            model.addAttribute("Message", msg);
+            return "message";
+        } else {
+            String msg = "Mail ID is already taken by someone!";
+            redirectAttributes.addFlashAttribute("message", msg);
+            return "redirect:/users/form?userId=" + refUserID;
+        }
+    }
 
-        return "message";
+    @GetMapping("/users/inactiveUsers")
+    public String view(Model model, HttpSession session) {
+        if (session.getAttribute("LoginUser") == null) {
+            return "redirect:/";
+
+        }
+        List<User> usersList = service.fetchAllUsers();
+        System.out.println(usersList);
+
+        model.addAttribute("UserList", usersList);
+
+        return "InactiveUsers";
 
     }
